+++ NEWS.md
@@ -7,6 +7,8 @@ Version 0.16.2 - dev
 
 - CHANGED: `ezdxf.path.add_bezier4p()`, add linear Bézier curve segments as LINE_TO commands
 - CHANGED: `ezdxf.path.add_bezier3p()`, add linear Bézier curve segments as LINE_TO commands
+- CHANGED: `$FINGERPRINTGUID` matches AutoCAD pattern `{XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX}`
+- CHANGED: `$VERSIONGUID` matches AutoCAD pattern `{XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX}`
 - BUGFIX: check for degenerated Bézier curves in `have_bezier_curves_g1_continuity()`
 
 Version 0.16.1 - 2021-04-10